// src/contracts/ERC20.schemas.ts
// Generated by abi-to-zod@0.2.0

import { z } from 'zod';

// === Primitive schemas ===

const ADDRESS = z
  .string()
  .regex(/^0x[0-9a-fA-F]{40}$/)
  .transform((v) => v as `0x${string}`);
const UINT256 = z
  .string()
  .regex(/^\d+$/)
  .transform((v) => BigInt(v))
  .refine((n) => n <= (1n << 256n) - 1n);

// === Function schemas ===

// allowance(address,address)
export const allowanceSchema = z.tuple([
  ADDRESS.describe('owner: address'),
  ADDRESS.describe('spender: address'),
]);

// approve(address,uint256)
export const approveSchema = z.tuple([
  ADDRESS.describe('spender: address'),
  UINT256.describe('amount: uint256'),
]);

// balanceOf(address)
export const balanceOfSchema = z.tuple([ADDRESS.describe('account: address')]);

// decimals()
export const decimalsSchema = z.tuple([]);

// decreaseAllowance(address,uint256)
export const decreaseAllowanceSchema = z.tuple([
  ADDRESS.describe('spender: address'),
  UINT256.describe('subtractedValue: uint256'),
]);

// increaseAllowance(address,uint256)
export const increaseAllowanceSchema = z.tuple([
  ADDRESS.describe('spender: address'),
  UINT256.describe('addedValue: uint256'),
]);

// name()
export const nameSchema = z.tuple([]);

// symbol()
export const symbolSchema = z.tuple([]);

// totalSupply()
export const totalSupplySchema = z.tuple([]);

// transfer(address,uint256)
export const transferSchema = z.tuple([
  ADDRESS.describe('recipient: address'),
  UINT256.describe('amount: uint256'),
]);

// transferFrom(address,address,uint256)
export const transferFromSchema = z.tuple([
  ADDRESS.describe('sender: address'),
  ADDRESS.describe('recipient: address'),
  UINT256.describe('amount: uint256'),
]);

// === Schema table ===

export const schemas = {
  'allowance': allowanceSchema,
  'approve': approveSchema,
  'balanceOf': balanceOfSchema,
  'decimals': decimalsSchema,
  'decreaseAllowance': decreaseAllowanceSchema,
  'increaseAllowance': increaseAllowanceSchema,
  'name': nameSchema,
  'symbol': symbolSchema,
  'totalSupply': totalSupplySchema,
  'transfer': transferSchema,
  'transferFrom': transferFromSchema,
  'allowance(address,address)': allowanceSchema,
  'approve(address,uint256)': approveSchema,
  'balanceOf(address)': balanceOfSchema,
  'decimals()': decimalsSchema,
  'decreaseAllowance(address,uint256)': decreaseAllowanceSchema,
  'increaseAllowance(address,uint256)': increaseAllowanceSchema,
  'name()': nameSchema,
  'symbol()': symbolSchema,
  'totalSupply()': totalSupplySchema,
  'transfer(address,uint256)': transferSchema,
  'transferFrom(address,address,uint256)': transferFromSchema,
} as const;